resource page: add initial loading mask
When having a lot of resources (or just a slow connection or server)
the load call can take several seconds.

Previously one just saw "no data loaded" with an error icon triangle
during the load, so no good feedback that the UI is just waiting on
the load to finish. So simply add a loading mask until the initial
load finished.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>