dashboard: add progress loading bar to each card
Similar to the previous commit but for the dashboard page. As we query
the cluster resources and node in two API requests and got three cards
there, were:
- the Analytics card depends on the result of both API calls
- the Nodes card depends on the result of just the nodes API call
- the Resources card depends on the result of just the resources call

So add two booleans tracking the loading of each separate API call
and use them according to above dependencies to decide when we should
show a dedicated loading bar for each of those cards.

This avoids having a ugly and confusing "no data loaded" error visible
during load, which looks like flickering if the API is fast (<1s) and
can be confusing if the API is slow (>1s).

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>